[change] - bookCar() method done

+++ backend-service/src/main/java/com/sda/rentalloapp/domain/CarBooking.java
@@ -29,7 +29,7 @@ public class CarBooking {
 
     LocalDate endDate;
 
-    int totalPriceInPolishGrosz;
+    Long totalPriceInPolishGrosz;
 
     @ManyToOne
     Address startLocation;
